Added new api.h util func and split io_server into more files
