+++ io_server/CMakeLists.txt
@@ -12,6 +12,12 @@ project(
 set(GENERATED_HEADERS ${PROJECT_BINARY_DIR}/generated_headers)
 configure_file(config/mainConfig.h.in "${GENERATED_HEADERS}/mainConfig.h")
 
+# Add wiringPi to the project 
+find_library( WIRINGPI_LIBRARIES wiringPi )
+find_path( WIRINGPI_INCLUDE_DIRS NAMES "wiringPi.h" )
+include_directories( ${WIRINGPI_INCLUDE_DIRS} )
+message(STATUS "wiringPi: lib:${WIRINGPI_LIBRARIES}, incl:${WIRINGPI_INCLUDE_DIRS}")
+
 # Add libnfc to the project
 find_library( LIBNFC_LIBRARIES nfc )
 find_path( LIBNFC_INCLUDE_DIR nfc )
@@ -22,11 +28,16 @@ message( STATUS "LibNFC: ${LIBNFC_INCLUDE_DIR}" )
 find_package( Boost REQUIRED COMPONENTS system thread  )
 include_directories( ${Boost_INCLUDE_DIR} )
 
-# Add wiringPi to the project 
-find_library( WIRINGPI_LIBRARIES wiringPi )
-find_path( WIRINGPI_INCLUDE_DIRS wiringPi )
-include_directories( ${WIRINGPI_INCLUDE_DIRS} )
-message(STATUS "wiringPi: lib:${WIRINGPI_LIBRARIES}, incl:${WIRINGPI_INCLUDE_DIRS}")
+# Add nlohmann_json::nlohmann_json
+include(FetchContent)
+FetchContent_Declare(json
+    GIT_REPOSITORY https://github.com/ArthurSonzogni/nlohmann_json_cmake_fetchcontent 
+    GIT_TAG v3.9.1)
+FetchContent_GetProperties(json)
+if(NOT json_POPULATED)
+    FetchContent_Populate(json)
+    add_subdirectory(${json_SOURCE_DIR} ${json_BINARY_DIR} EXCLUDE_FROM_ALL)
+endif()
 
 # The compiled library code is here
 add_subdirectory(src)
@@ -47,5 +58,5 @@ if(CMAKE_PROJECT_NAME STREQUAL PROJECT_NAME)
 
   include_directories(${GENERATED_HEADERS})
 
-  target_link_libraries(io_server PRIVATE ioserver ${Boost_LIBRARIES} ${WIRINGPI_LIBRARIES} ${LIBNFC_LIBRARIES})
+  target_link_libraries(io_server PRIVATE ioserver ${Boost_LIBRARIES} ${WIRINGPI_LIBRARIES} ${LIBNFC_LIBRARIES} nlohmann_json::nlohmann_json)
 endif()
